moved sspconfig to types
